requirements.txt
fastapi
uvicorn
opencv-python
numpy
pydantic
rknn-toolkit2
python-multipart
